cleaned up main a little. Still getting compilation errors due to a circular inclusion problem with map and tile
