+++ src/utility.cc
@@ -39,20 +39,3 @@ sf::ConvexShape generateHex(double x, double y, double r) {
 
     return hex;
 }
-
-sf::Vector2f AxialToPixel(const AxialCoordinate& p_coordinate) {
-    double x = HEX_RADIUS * (1.5 * p_coordinate.q);
-    double y = HEX_RADIUS * (((ROOT3 / 2) * p_coordinate.q) + (ROOT3 * p_coordinate.r));
-    return sf::Vector2f(x, y);
-}
-sf::Vector2f AxialToPixel(const AxialCoordinate&& p_coordinate) {
-    double x = HEX_RADIUS * (1.5 * p_coordinate.q);
-    double y = HEX_RADIUS * (((ROOT3 / 2) * p_coordinate.q) + (ROOT3 * p_coordinate.r));
-    return sf::Vector2f(x, y);
-}
-
-AxialCoordinate PixelToAxial(size_t x, size_t y) {
-    double q = (2/3 * double(x)) / HEX_RADIUS;
-    double r = (-1/3 * double(x) + ROOT3/3 * double(y)) / HEX_RADIUS;
-    return AxialCoordinate(q, r);
-}
